Add Tkinter interface for file selection and clip range
